Use Unicode Blocks instead of Scripts
Huge Update: There is a list of Unicode Blocks, so we can use that to
separate the characters into files. Thus, some characters are shuffled
around, but most importantly, a lot of stuff has been added!

While doing this, the extractors have also been consolidated so that the
character names are downloaded once, for example.

The `EmojiExtractor` still lives on, as it also supports joined emoji,
which are contained in the basic Blocks.
Similarly, the `MathExtractor` collects a lot of characters that are in
different blocks. So for now, it stays.