Support running as a rofi mode
A much requested feature was to be able to run `rofimoji` as a rofi
"mode". With this, you could, for example, have a combi mode with
different character sets.

However, rofi unfortunately (and sometimes understandably) limits the
features of its modes. The most annoying is that you cannot call
`xdotool` to type into another window. So with the mode, the best you
can do is to copy to the clipboard.
Secondly, all the custom shortcut mappings are gone, and we must use the
default ones which are not easy to remember.
Finally, selecting more than one character is also impossible.

Is it worth it? I'm not sure, but we can try.

Issue: #44