Initial support for wofi
Just like the `Typer` and `Clipboarder`, the actual call to `rofi` is
now abstracted away behind a facade, and that allows us to also use
`wofi`.

Unfortunately, `wofi` does not support some cool features like recent
characters or custom keyboard shortcuts. In these cases, the user can't
use it but `rofimoji` would be ready to do it - the code is already
there.