Add ltr to all characters
Even rigth-to-left languages like Arabic or Hebrew are easier to read
with their English description if they're formatted left-to-right with
the character first.
If we ever have descriptions in other languages, we must re-evaluate
that argument, but for now, I think it's better like this.

Also, more classes now use the `Character` type. That should hopefully
simplify something, too.