Use Unicode data for directionality, if possible
We already downloaded the relevant data from the official source, so we
should prefer it over Python's built in information. THis is both
because Python's might be outdated, and because we don't want to be
coupled to a certain Python version.

However, for Nerdfonts (and maybe others, in the future), we do not have
this information. In these cases, we need a fallback solution, and that
can be Python.