Stop rofi from re-sorting
Letting rofi re-sort the entries is nice for searching new emojis.
However, it completely disregards frecency, so entries that you often
use may be moved downwards in favor of things that might be a better
match but never used. This is probably not ideal when one assumes that
some entries are used often for a reason.

Issue: #110