Split it up
where 'it' means everything 🙈
rofimoji got too large for just one file, so:
- common functionality is split into separate, much smaller files with
  just one topic
- typer, selector and clipboarder are now separate modules
- standalone rofimoji and rofi mode rofimoji are finally separated
- it's possible to understand the rofi mode again 🪅

Especially the separate rofi mode needed lots of changes to make it easy
to understand and not just an afterthought. There might be some smaller
breakages, but overall it looks so much better and it seems to work (so
far).