Split recents by fileset
So far, the recently chosen characters were shared among all instances
of `rofimoji`, no matter which files were actually shown.

Now, we show and save them per fileset, which is an improvement.
However, if one file is used in multiple filesets, their usage is not
shared, but separate. A solution for this seems to complicated for what
we try to achieve here.

The old `recents` file will be migrated on startup, if necessary.

Many thanks to @will-lynas for reporting this and working with me to
find a good solution.

Issue: #173