Removed hard dependency on header files
Changes implemented by adding a new method to actions.py which gets
the "event code" of a key. This method is used in the ydotool.py
script to get the event code of a key and then use it to send the
key press event to the system.
Has a helper dictionary to map key names to event codes in
input_event_codes.py

Also I cleaned up some code; removed the socket thing.